fixes bob fixes the size of the buffer fixes shifting, which might be incorrect if msb is 1 (negative) adds more debug output renames constants adds code to print data better uses larger data types to prevent under/overflow

+++ Code/include/espHelper/src/espHelper.c
@@ -148,10 +148,11 @@ void initEsp() {
  * @param numberOfBytes 
  * @param CIP_Command 
  */
-void sendCip(const size_t numberOfBytes, char* command) {
+void sendCip(const uint64_t numberOfBytes, char* command) {
     strcpy(command, "AT+CIPSEND=");
     char charBytesToSend[80];
     sprintf(charBytesToSend, "%llu", numberOfBytes);
+    // printf("bytes being sent is: %s\n", charBytesToSend);
     strcat(command, charBytesToSend);
     strcat(command, "\r\n");
 }
